fixtures::json
Struct JsonFixture

pub struct JsonFixture { /* private fields */ }

Implementations

impl JsonFixture

pub fn all_fixtures() -> &'static BTreeMap<&'static str, JsonFixture>

pub fn metadata1() -> &'static Self

pub fn metadata2() -> &'static Self

pub fn metadata_builddep() -> &'static Self

pub fn metadata_dups() -> &'static Self

pub fn metadata_cycle1() -> &'static Self

pub fn metadata_cycle2() -> &'static Self

pub fn metadata_cycle_features() -> &'static Self

pub fn metadata_targets1() -> &'static Self

pub fn metadata_build_targets1() -> &'static Self

pub fn metadata_proc_macro1() -> &'static Self

pub fn metadata_alternate_registries() -> &'static Self

pub fn metadata_weak_namespaced_features() -> &'static Self

pub fn metadata_libra() -> &'static Self

pub fn metadata_libra_f0091a4() -> &'static Self

pub fn metadata_libra_9ffd93b() -> &'static Self

pub fn mnemos_b3b4da9() -> &'static Self

pub fn hyper_util_7afb1ed() -> &'static Self

pub fn metadata_guppy_78cb7e8() -> &'static Self

pub fn metadata_guppy_869476c() -> &'static Self

pub fn metadata_guppy_c9b4f76() -> &'static Self

pub fn metadata_guppy_44b62fa() -> &'static Self

impl JsonFixture

pub fn by_name(name: &str) -> Option<&'static Self>

Lookup a fixture by name, or None if the name wasn’t found.

pub fn name(&self) -> &'static str

Returns the name of this fixture.

pub fn abs_path(&self) -> &Utf8Path

Returns the absolute path of this fixture.

pub fn workspace_path(&self) -> &Utf8Path

Returns the path of this fixture, relative to the workspace root.

pub fn json(&self) -> &str

Returns the unparsed JSON string for this fixture.

pub fn graph(&self) -> &PackageGraph

Returns the package graph for this fixture.

pub fn details(&self) -> &FixtureDetails

Returns the test details for this fixture.

pub fn verify(&self)

Verifies that the parsed metadata matches known details.