Module fixtures::dep_helpers

Structs

DirectionDesc

Some of the messages are different based on whether we’re testing forward deps or reverse ones. For forward deps, we use the terms “known” for ‘from’ and “variable” for ‘to’. For reverse deps it’s the other way round.

Traits

GraphAssert

GraphMetadata

GraphQuery

GraphSet

Functions

assert_link_order

Assert that links are presented in the expected order.