fixtures::details
Struct FixtureDetails

pub struct FixtureDetails { /* private fields */ }

This captures metadata fields that are relevant for tests. They are meant to be written out lazily as tests are filled out – feel free to add more details as necessary!

Implementations

impl FixtureDetails

pub fn new(package_details: AHashMap<PackageId, PackageDetails>) -> Self

pub fn with_workspace_members<'a>( self, workspace_members: impl IntoIterator<Item = (impl Into<Utf8PathBuf>, &'a str)>, ) -> Self

pub fn with_link_details( self, link_details: AHashMap<(PackageId, PackageId), LinkDetails>, ) -> Self

pub fn with_feature_graph_warnings( self, warnings: Vec<FeatureGraphWarning>, ) -> Self

pub fn with_cycles(self, cycles: Vec<Vec<&'static str>>) -> Self

pub fn known_ids(&self) -> impl Iterator<Item = &PackageId>

pub fn assert_workspace(&self, workspace: Workspace<'_>)

pub fn assert_topo(&self, graph: &PackageGraph)

pub fn assert_metadata( &self, id: &PackageId, metadata: PackageMetadata<'_>, msg: &str, )

pub fn has_build_targets(&self, id: &PackageId) -> bool

pub fn assert_build_targets(&self, metadata: PackageMetadata<'_>, msg: &str)

pub fn has_deps(&self, id: &PackageId) -> bool

Returns true if the deps for this package are available to test against.

pub fn assert_deps(&self, graph: &PackageGraph, id: &PackageId, msg: &str)

pub fn has_reverse_deps(&self, id: &PackageId) -> bool

Returns true if the reverse deps for this package are available to test against.

pub fn assert_reverse_deps( &self, graph: &PackageGraph, id: &PackageId, msg: &str, )

pub fn has_transitive_deps(&self, id: &PackageId) -> bool

Returns true if the transitive deps for this package are available to test against.

pub fn assert_transitive_deps( &self, graph: &PackageGraph, id: &PackageId, msg: &str, )

pub fn has_transitive_reverse_deps(&self, id: &PackageId) -> bool

Returns true if the transitive reverse deps for this package are available to test against.

pub fn assert_transitive_reverse_deps( &self, graph: &PackageGraph, id: &PackageId, msg: &str, )

pub fn assert_link_details(&self, graph: &PackageGraph, msg: &str)

pub fn has_named_features(&self, id: &PackageId) -> bool

pub fn assert_named_features( &self, graph: &PackageGraph, id: &PackageId, msg: &str, )

pub fn assert_feature_graph_warnings(&self, graph: &PackageGraph, msg: &str)

pub fn assert_cycles(&self, graph: &PackageGraph, msg: &str)